Add files via upload
These are the same version of the ANATHING-BASE_OSH_V12_production and ANATHING-FRONT_OSH_V12a_production except that there is an additional BoardOutline layer added. The board outlines from all other gerber are removed and there is a little drill hole change for the 4 outer mechanical 3.2mm holes for the screws (which have been as milled structure in the version before). Please be aware, that the ANATHING-FRONT BoardOutline layer does not only contain the outline but also milling structures for the big holes for the potentiometers, switches and the panelmeter. There can not be devided further by the software DesignSpark PCB Pro, as it is usually handled in the same process when milling the outline it is the same production step to mill the "inlines". This step can be done by the pcb manufacturer if needed. If you pcb manufacturer is not able to handle this, you may consider to change you manufacturing company.